python3-requirements.txt
Pillow==7.1.1
pkg-resources==0.0.0
python-xlib==0.27
six==1.14.0
